requirements.txt
asgiref==3.5.2
Django==4.1.4
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
